# backend/internal/contractassets/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zenbo Agent Gateway",
    "version": "0.1.0",
    "description": "HTTP and native WebSocket contract for a Zenbo-K thin client. Provider credentials remain exclusively in the external gateway."
  },
  "servers": [
    {
      "url": "https://gateway.example.com/agent/v1",
      "description": "Deployment-specific Agent Gateway base URL"
    }
  ],
  "security": [
    {
      "DeviceBearer": []
    }
  ],
  "paths": {
    "/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "summary": "Read gateway protocol and media limits",
        "parameters": [
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" }
        ],
        "responses": {
          "200": {
            "description": "Gateway capabilities for protocol 1.0",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Capabilities" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      }
    },
    "/sessions": {
      "post": {
        "operationId": "createSession",
        "summary": "Create one active Agent Gateway session for the device",
        "parameters": [
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSessionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created or the idempotent prior result returned",
            "headers": {
              "Location": {
                "required": true,
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/sessions/{sessionId}": {
      "parameters": [
        { "$ref": "#/components/parameters/SessionId" }
      ],
      "get": {
        "operationId": "getSession",
        "summary": "Read session state and replay cursor",
        "parameters": [
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" }
        ],
        "responses": {
          "200": {
            "description": "Current session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      },
      "delete": {
        "operationId": "deleteSession",
        "summary": "Close a session idempotently",
        "parameters": [
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" }
        ],
        "responses": {
          "204": { "description": "Session is closed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      }
    },
    "/sessions/{sessionId}/turns": {
      "post": {
        "operationId": "createTurn",
        "summary": "Submit one completed text or voice turn",
        "description": "Accepts either JSON text or mono PCM WAV audio up to 30 seconds and 2 MiB. The gateway performs agent orchestration asynchronously.",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TextTurn" }
            },
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/TurnUpload" },
              "encoding": {
                "audio": {
                  "contentType": "audio/wav"
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Turn accepted for asynchronous processing",
            "headers": {
              "Location": {
                "required": true,
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TurnAccepted" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "422": { "$ref": "#/components/responses/Unprocessable" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/sessions/{sessionId}/turns/{turnId}/cancel": {
      "post": {
        "operationId": "cancelTurn",
        "summary": "Request cancellation of an active turn",
        "description": "Cancellation is idempotent. A terminal turn remains terminal and is returned unchanged.",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/TurnId" },
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CancelTurnRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Cancellation accepted or turn already terminal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TurnState" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      }
    },
    "/sessions/{sessionId}/tool-calls/{callId}": {
      "put": {
        "operationId": "putToolCallResult",
        "summary": "Report a correlated device tool-call state or terminal result",
        "description": "Updates are monotonic. Repeating an identical representation is idempotent; attempting to replace a terminal result returns 409.",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/CallId" },
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ToolCallUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored tool-call state",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ToolCallUpdate" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "422": { "$ref": "#/components/responses/Unprocessable" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      }
    },
    "/sessions/{sessionId}/audio/{artifactId}": {
      "get": {
        "operationId": "getAudioArtifact",
        "summary": "Download one gateway-produced audio artifact",
        "description": "The response body is limited to 10 MiB and its SHA-256 digest must match the artifact metadata received over the event stream.",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/ArtifactId" },
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" }
        ],
        "responses": {
          "200": {
            "description": "Audio bytes; at most 10 MiB",
            "headers": {
              "Digest": {
                "required": true,
                "description": "RFC 9530 SHA-256 content digest",
                "schema": { "type": "string" }
              },
              "Content-Length": {
                "required": true,
                "schema": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10485760
                }
              }
            },
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "contentEncoding": "binary"
                }
              },
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "contentEncoding": "binary"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "410": { "$ref": "#/components/responses/Gone" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      }
    },
    "/sessions/{sessionId}/playback": {
      "post": {
        "operationId": "reportPlayback",
        "summary": "Report device playback lifecycle for a requested audio artifact",
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlaybackUpdate" }
            }
          }
        },
        "responses": {
          "202": { "description": "Playback update accepted" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      }
    },
    "/sessions/{sessionId}/events": {
      "get": {
        "operationId": "connectEventStream",
        "summary": "Upgrade to the outbound session WebSocket event stream",
        "description": "Native OkHttp must send the same Authorization bearer token and Zenbo headers as HTTP. The stream is gateway-to-device; reconnect with after equal to the last durably handled sequence.",
        "x-websocket": true,
        "x-message-schema": "./schemas/ws-envelope.schema.json",
        "x-stale-cursor-recovery": {
          "eventType": "session.snapshot",
          "cursorField": "data.lastSequence"
        },
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" },
          { "$ref": "#/components/parameters/DeviceId" },
          { "$ref": "#/components/parameters/ProtocolVersion" },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Last sequence durably handled by the device; the server replays events with a greater sequence. Omit on first connection.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "101": { "description": "WebSocket upgrade accepted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "The requested replay cursor is ahead of the session cursor",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "426": { "$ref": "#/components/responses/ProtocolMismatch" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "DeviceBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque device-scoped token",
        "description": "One revocable token scoped to a single Zenbo device and the Agent Gateway API. It is not a provider credential."
      }
    },
    "parameters": {
      "DeviceId": {
        "name": "X-Zenbo-Device-Id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
      "ProtocolVersion": {
        "name": "X-Zenbo-Protocol",
        "in": "header",
        "required": true,
        "schema": { "const": "1.0" }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Client-generated UUID retained for at least the session lifetime.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, invalid, revoked, or device-mismatched credential",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "ProtocolMismatch": {
        "$ref": "#/components/responses/ErrorResponse"
      },
...
    },
    "schemas": {
      "Capabilities": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "protocolVersion",
          "audioInput",
          "audioOutput",
          "eventTypes",
          "agentProfiles",
          "retentionPolicy"
        ],
        "properties": {
          "protocolVersion": { "const": "1.0" },
          "audioInput": {
            "type": "object",
            "additionalProperties": false,
            "required": ["contentTypes", "maxBytes", "maxDurationMs"],
            "properties": {
              "contentTypes": {
                "type": "array",
                "prefixItems": [{ "const": "audio/wav" }],
                "items": false,
                "minItems": 1,
                "maxItems": 1
              },
              "maxBytes": { "const": 2097152 },
              "maxDurationMs": { "const": 30000 }
            }
          },
          "audioOutput": {
            "type": "object",
            "additionalProperties": false,
            "required": ["contentTypes", "maxBytes"],
            "properties": {
              "contentTypes": {
                "type": "array",
                "items": { "enum": ["audio/mpeg", "audio/wav"] },
                "minItems": 1,
                "uniqueItems": true
              },
              "maxBytes": { "const": 10485760 }
            }
          },
          "eventTypes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GatewayEventType" },
            "minItems": 1,
            "uniqueItems": true
          },
          "agentProfiles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AgentProfile" },
            "minItems": 1
          },
          "retentionPolicy": {
            "$ref": "#/components/schemas/RetentionPolicy"
          }
        }
      },
...
    }
  }
}